# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: repcal.info
  version: '2.0'
  license:
    name: MIT
    url: 'https://spdx.org/licenses/MIT.html'
  contact:
    name: Tomas Thelander
    url: 'https://tthe.se'
    email: tomas@tthe.se
  description: |
    Welcome to the API for Repcal.info.
    
    As this is a hypermedia API, it can be explored by starting at the [Index resource](/operations/get-api).
servers:
  - url: 'https://repcal.info'
tags:
  - name: Resources
    description: Resources
paths:
  /api:
    get:
      summary: API Index
      description: |
        This resource serves as a starting point for exploring the API.
        
        The impatient can go directly finding out [the current time](/operations/get-now).
      responses:
        '200':
          description: ''
          headers: {}
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ApiIndex'
              example:
                _links: {}
                name: ''
                description: ''
      operationId: get-api
  /now:
    get:
      summary: Get current time
      operationId: get-now
      parameters:
        - schema:
            type: number
            exclusiveMinimum: -1440
            exclusiveMaximum: 1440
            example: 60

          in: query
          name: offset
          description: Specify offset as a positive or negative number of minutes from UTC.
      description: Operation that will resolve and redirect to the Moment resource for the current date and time.
      responses:
        '302':
          description: ''
          headers:
            Location:
              schema:
                type: string
                example: /moment/12345/0
              description: The location of the resolved resource.
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/moment/{timestamp}/{offset}':
    parameters:
      - schema:
          type: integer
        name: timestamp
        in: path
        required: true
        description: UNIX timestamp
      - schema:
          type: number
          exclusiveMinimum: -1440
          exclusiveMaximum: 1440
          example: 60
        name: offset
        in: path
        required: true
        description: Number of minutes from UTC
    get:
      summary: Moment
      tags:
        - Resources
      responses:
        '200':
          $ref: '#/components/responses/MomentResponse'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      operationId: get-moment
      description: |
        The Moment resource represents a specific point in time.
        
        The HAL response by the API will embed the related republican-style Date and Time resources.
  '/date/{year}/{month}/{day}':
    parameters:
      - schema:
          type: integer
          example: 2023
        name: year
        in: path
        required: true
        description: Year (Gregorian)
      - schema:
          type: integer
          example: 2
        name: month
        in: path
        required: true
        description: Month (Gregorian)
      - schema:
          type: integer
          example: 12
        name: day
        in: path
        required: true
        description: Day (Gregorian)
    get:
      summary: Date
      description: The Date resource contains information about a given date in the republican calendar.
      tags:
        - Resources
      responses:
        '200':
          description: ''
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/RepublicanDate'
              example:
                _links: {}
                texts:
                  default: 'Sextidi 26 germinal an CCXXXI'
                attributes: {}
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      operationId: get-date
  '/time/{hour}/{minute}/{second}':
    parameters:
      - schema:
          type: integer
          example: 20
        name: hour
        in: path
        required: true
        description: Hour (standard)
      - schema:
          type: integer
          example: 49
        name: minute
        in: path
        required: true
        description: Minute (standard)
      - schema:
          type: integer
          example: 12
        name: second
        in: path
        required: true
        description: Second (standard)
    get:
      summary: Time
      description: The Time resource contains information about a time of day in a decimal format.
      tags:
        - Resources
      responses:
        '200':
          description: ''
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/RepublicanTime'
              example:
                _links: {}
                texts:
                  default: '8:99:40'
                  decimal: '0,8994'
                attributes: {}
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      operationId: get-time
components:
  schemas:
    ApiIndex:
      $ref: '/meta/schema/api'
    Moment:
      $ref: '/meta/schema/moment'
    RepublicanTime:
      $ref: '/meta/schema/time'
    RepublicanDate:
      $ref: '/meta/schema/date'
    ProblemDetails:
      title: Problem Details
      type: object
      description: |
        Problem details representing an error that has occurred.
        Follows the [RFC 7807](https://datatracker.ietf.org/doc/html/rfc7807) format.
      examples:
        - type: 'https://webconcepts.info/concepts/http-status-code/400'
          instance: 'https://repcal.info/meta/errors/400/2cf1c733-90f2-49d2-a700-5be04efc3ddd'
          status: 400
          title: Bad Request
          detail: The provided offset was invalid or out-of-bounds.
      x-stoplight:
        id: a39e635591f39
      properties:
        type:
          type: string
          format: uri
        instance:
          type: string
          format: uri
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
      required:
        - type
        - instance
        - status
        - title
        - detail
  responses:
    MomentResponse:
      description: ''
      content:
        application/hal+json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Moment'
              - type: object
                properties:
                  _embedded:
                    type: object
                    properties:
                      'rel:date':
                        $ref: '#/components/schemas/RepublicanDate'
                      'rel:time':
                        $ref: '#/components/schemas/RepublicanTime'
          example:
            _links: {}
            iso: '2023-04-15T21:35:09+0200'
            _embedded:
              repcal:date: {}
              repcal:time: {}